requirements/base.txt
Django>=4.0,<4.1
wagtail>=3.0,<3.1
django-manifest-loader
djangorestframework
